# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GLOWNDRY Laundry Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CustomerRegistration {
  /** @minLength 2 */
  fullName: string;
  email: string;
  phone: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginCredentials {
  email: string;
  password: string;
}

export interface Customer {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  totalOrders?: number;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  customer: Customer;
}

export interface CustomerUpdate {
  fullName?: string;
  phone?: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  pricePerKg: number;
  createdAt: string;
}

export interface ServiceInput {
  name: string;
  description: string;
  pricePerKg: number;
}

export interface ServiceUpdate {
  name?: string;
  description?: string;
  pricePerKg?: number;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  scheduled: 'scheduled',
  received: 'received',
  in_progress: 'in_progress',
  ready: 'ready',
  claimed: 'claimed',
} as const;

export interface Booking {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  serviceId: number;
  /** @nullable */
  serviceName?: string | null;
  scheduledDate: string;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  totalAmount: number;
  createdAt: string;
}

export type BookingDetailStatus = typeof BookingDetailStatus[keyof typeof BookingDetailStatus];


export const BookingDetailStatus = {
  scheduled: 'scheduled',
  received: 'received',
  in_progress: 'in_progress',
  ready: 'ready',
  claimed: 'claimed',
} as const;

export interface BookingItem {
  id: number;
  bookingId: number;
  clothingType: string;
  quantity: number;
  /** @nullable */
  notes?: string | null;
}

export interface BookingDetail {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  serviceId: number;
  /** @nullable */
  serviceName?: string | null;
  scheduledDate: string;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  notes?: string | null;
  status: BookingDetailStatus;
  totalAmount: number;
  createdAt: string;
  items: BookingItem[];
}

export interface BookingInput {
  customerId: number;
  serviceId: number;
  scheduledDate: string;
  weightKg?: number;
  notes?: string;
}

export type BookingStatusUpdateStatus = typeof BookingStatusUpdateStatus[keyof typeof BookingStatusUpdateStatus];


export const BookingStatusUpdateStatus = {
  scheduled: 'scheduled',
  received: 'received',
  in_progress: 'in_progress',
  ready: 'ready',
  claimed: 'claimed',
} as const;

export interface BookingStatusUpdate {
  status: BookingStatusUpdateStatus;
  weightKg?: number;
  totalAmount?: number;
}

export interface BookingItemInput {
  clothingType: string;
  quantity: number;
  notes?: string;
}

export interface BookingItemUpdate {
  clothingType?: string;
  quantity?: number;
  notes?: string;
}

export interface Payment {
  id: number;
  bookingId: number;
  /** @nullable */
  customerName?: string | null;
  amount: number;
  cashReceived: number;
  change: number;
  createdAt: string;
}

export interface PaymentInput {
  bookingId: number;
  amount: number;
  cashReceived: number;
}

export interface Notification {
  id: number;
  customerId: number;
  title: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface NotificationInput {
  customerId: number;
  title: string;
  message: string;
}

export interface DashboardStats {
  activeOrders: number;
  dailyRevenue: number;
  expectedDropoffs: number;
  totalCustomers: number;
  totalBookings: number;
  recentBookings?: Booking[];
}

export type ListCustomersParams = {
search?: string;
};

export type ListBookingsParams = {
customerId?: number;
status?: string;
search?: string;
};

export type ListNotificationsParams = {
customerId: number;
};