# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: GLOWNDRY Laundry Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerCustomer
      tags: [auth]
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerRegistration"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: loginCustomer
      tags: [auth]
      summary: Customer login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/admin/login:
    post:
      operationId: loginAdmin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Admin logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Customers ─────────────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers (admin)
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update customer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"

  # ── Services ──────────────────────────────────────────────────────────────────
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bookings list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a booking (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get booking by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      summary: Update booking status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingStatusUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ── Booking Items ─────────────────────────────────────────────────────────────
  /bookings/{id}/items:
    get:
      operationId: listBookingItems
      tags: [items]
      summary: List items for a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookingItem"
    post:
      operationId: addBookingItem
      tags: [items]
      summary: Add item to a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingItemInput"
      responses:
        "201":
          description: Created item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingItem"
  /items/{id}:
    patch:
      operationId: updateBookingItem
      tags: [items]
      summary: Update a booking item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingItemUpdate"
      responses:
        "200":
          description: Updated item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingItem"
    delete:
      operationId: deleteBookingItem
      tags: [items]
      summary: Delete a booking item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Payments ──────────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for a customer
      parameters:
        - name: customerId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notifications list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
    post:
      operationId: createNotification
      tags: [notifications]
      summary: Create a notification (admin/system)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"

  # ── Dashboard ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ────────────────────────────────────────────────────────────────────
    CustomerRegistration:
      type: object
      required: [fullName, email, phone, password]
      properties:
        fullName:
          type: string
          minLength: 2
        email:
          type: string
          format: email
        phone:
          type: string
        password:
          type: string
          minLength: 6
    LoginCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, customer]
      properties:
        token:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"

    # ── Customer ─────────────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, fullName, email, phone, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        totalOrders:
          type: integer
        createdAt:
          type: string
          format: date-time
    CustomerUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: string

    # ── Service ──────────────────────────────────────────────────────────────────
...
    Booking:
      type: object
      required: [id, customerId, serviceId, scheduledDate, status, totalAmount, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        serviceId:
          type: integer
        serviceName:
          type: ["string", "null"]
        scheduledDate:
          type: string
          format: date-time
        weightKg:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, received, in_progress, ready, claimed]
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time
    BookingDetail:
      type: object
      required: [id, customerId, serviceId, scheduledDate, status, totalAmount, createdAt, items]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        serviceId:
          type: integer
        serviceName:
          type: ["string", "null"]
        scheduledDate:
          type: string
          format: date-time
        weightKg:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        status:
          type: string
          enum: [scheduled, received, in_progress, ready, claimed]
        totalAmount:
          type: number
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/BookingItem"
    BookingInput:
      type: object
      required: [customerId, serviceId, scheduledDate]
      properties:
        customerId:
          type: integer
        serviceId:
          type: integer
        scheduledDate:
          type: string
          format: date-time
        weightKg:
          type: number
        notes:
          type: string
    BookingStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [scheduled, received, in_progress, ready, claimed]
        weightKg:
          type: number
        totalAmount:
          type: number

    # ── Booking Item ─────────────────────────────────────────────────────────────
    BookingItem:
      type: object
      required: [id, bookingId, clothingType, quantity]
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        clothingType:
          type: string
        quantity:
          type: integer
        notes:
          type: ["string", "null"]
    BookingItemInput:
      type: object
      required: [clothingType, quantity]
      properties:
        clothingType:
          type: string
        quantity:
          type: integer
        notes:
          type: string
    BookingItemUpdate:
      type: object
      properties:
        clothingType:
          type: string
        quantity:
          type: integer
        notes:
          type: string

    # ── Payment ──────────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, customerId, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
    NotificationInput:
      type: object
      required: [customerId, title, message]
      properties:
        customerId:
          type: integer
        title:
          type: string
        message:
          type: string

    # ── Dashboard ─────────────────────────────────────────────────────────────────
...