# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GLOWNDRY Laundry Management API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  serviceId: number;
  /** @nullable */
  serviceName?: string | null;
  scheduledDate: Date;
  /** @nullable */
  weightKg?: number | null;
  /** @nullable */
  notes?: string | null;
  status: BookingStatus;
  totalAmount: number;
  createdAt: Date;
}