# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * GLOWNDRY Laundry Management API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  customerId: number;
  serviceId: number;
  scheduledDate: Date;
  weightKg?: number;
  notes?: string;
}